+++ src/relaxations/cardinality_preserving.h
@@ -9,6 +9,6 @@ struct relaxation_solution solve_cardinality_preserving_relaxation(const struct 
 
 // Solves the cardinality-preserving relaxation, when we are allowed to take at most
 // maxb black and maxw white items. Vector ids contains items' indeces sorted by p/w
-struct relaxation_solution solve_cardinality_preserving_relaxation_for_bounds(const struct instance* inst, uint_fast32_t maxb, uint_fast32_t maxw, const ptrdiff_t* ids);
+struct relaxation_solution solve_cardinality_preserving_relaxation_for_bounds(const struct instance* inst, uint_fast64_t maxb, uint_fast64_t maxw, const ptrdiff_t* ids);
 
 #endif
